test: consolidate CLI subprocess fixture
